Update index.html
The entire animation is made of a number of smaller animations.  Each animation is allowed a set time to play.  The entire set of animations is allowed a set time to play.  The single animation time exceeds the entire animation playtime alotted.  Since a count of the number of animations is kept, this problem can be solved by increasing the allowable time for the entire animation set to play.  1000 -> 2000.